renamed temp icon
